+++ language-server/modules/langserver-core/src/test/java/org/ballerinalang/langserver/performance/CodeActionPerformanceTest.java
@@ -17,10 +17,7 @@
  */
 package org.ballerinalang.langserver.performance;
 
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import org.ballerinalang.langserver.codeaction.CodeActionUtil;
 import org.ballerinalang.langserver.common.utils.CommonUtil;
 import org.ballerinalang.langserver.commons.LanguageServerContext;
@@ -46,8 +43,6 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
@@ -55,7 +50,6 @@ import java.util.stream.Collectors;
  */
 public class CodeActionPerformanceTest {
     private Endpoint serviceEndpoint;
-    private final JsonParser parser = new JsonParser();
     private final Path testRoot = FileUtils.RES_DIR.resolve("performance");
     private static final WorkspaceManager workspaceManager
             = BallerinaWorkspaceManager.getInstance(new LanguageServerContextImpl());
@@ -88,80 +82,12 @@ public class CodeActionPerformanceTest {
         long start = System.currentTimeMillis();
         String res = TestUtil.getCodeActionResponse(serviceEndpoint, sourcePath.toString(), range, codeActionContext);
         long end = System.currentTimeMillis();
-        long responseTime = end - start;
-        Assert.assertEquals(responseTime < 3000, true);
+        long actualResponseTime = end - start;
+        int expectedResponseTime = Integer.parseInt(System.getProperty("responseTimeThreshold"));
+        Assert.assertTrue(actualResponseTime < expectedResponseTime,
+                String.format("Expected response time = %d, received %d.", expectedResponseTime, actualResponseTime));
 
-        for (JsonElement element : configJsonObject.get("expected").getAsJsonArray()) {
-            JsonObject expected = element.getAsJsonObject();
-            String expTitle = expected.get("title").getAsString();
-
-            boolean codeActionFound = false;
-            JsonObject responseJson = this.getResponseJson(res);
-            for (JsonElement jsonElement : responseJson.getAsJsonArray("result")) {
-                JsonObject right = jsonElement.getAsJsonObject().get("right").getAsJsonObject();
-                if (right == null) {
-                    continue;
-                }
-
-                // Match title
-                String actualTitle = right.get("title").getAsString();
-                if (!expTitle.equals(actualTitle)) {
-                    continue;
-                }
-                // Match edits
-                if (expected.get("edits") != null) {
-                    JsonArray actualEdit = right.get("edit").getAsJsonObject().get("documentChanges")
-                            .getAsJsonArray().get(0).getAsJsonObject().get("edits").getAsJsonArray();
-                    JsonArray expEdit = expected.get("edits").getAsJsonArray();
-                    if (!expEdit.equals(actualEdit)) {
-                        continue;
-                    }
-                }
-                // Match args
-                if (expected.get("command") != null) {
-                    JsonObject expectedCommand = expected.get("command").getAsJsonObject();
-                    JsonObject actualCommand = right.get("command").getAsJsonObject();
-
-                    if (!Objects.equals(actualCommand.get("command"), expectedCommand.get("command"))) {
-                        continue;
-                    }
-
-                    if (!Objects.equals(actualCommand.get("title"), expectedCommand.get("title"))) {
-                        continue;
-                    }
-
-                    JsonArray actualArgs = actualCommand.getAsJsonArray("arguments");
-                    JsonArray expArgs = expectedCommand.getAsJsonArray("arguments");
-                    if (!TestUtil.isArgumentsSubArray(actualArgs, expArgs)) {
-                        continue;
-                    }
-
-                    boolean docUriFound = false;
-                    for (JsonElement actualArg : actualArgs) {
-                        JsonObject arg = actualArg.getAsJsonObject();
-                        if ("doc.uri".equals(arg.get("key").getAsString())) {
-                            Optional<Path> docPath = CommonUtil.getPathFromURI(arg.get("value").getAsString());
-                            if (docPath.isPresent()) {
-                                // We just check file names, since one refers to file in build/ while
-                                // the other refers to the file in test resources
-                                docUriFound = docPath.get().getFileName().equals(sourcePath.getFileName());
-                            }
-                        }
-                    }
-
-                    if (!docUriFound) {
-                        continue;
-                    }
-                }
-                // Code-action matched
-                codeActionFound = true;
-                break;
-            }
-            String cursorStr = range.getStart().getLine() + ":" + range.getEnd().getCharacter();
-            Assert.assertTrue(codeActionFound,
-                    "Cannot find expected Code Action for: " + expTitle + ", cursor at " + cursorStr
-                            + " in " + sourcePath);
-        }
+        TestUtil.evaluateCodeActionTest(res, configJsonObject, sourcePath, range);
         TestUtil.closeDocument(this.serviceEndpoint, sourcePath);
     }
 
@@ -173,12 +99,6 @@ public class CodeActionPerformanceTest {
     @AfterClass
     public void cleanupLanguageServer() {
         TestUtil.shutdownLanguageServer(this.serviceEndpoint);
-    }
-
-    private JsonObject getResponseJson(String response) {
-        JsonObject responseJson = parser.parse(response).getAsJsonObject();
-        responseJson.remove("id");
-        return responseJson;
     }
 
     @DataProvider(name = "performance-data-provider")
