+++ language-server/modules/langserver-core/src/test/resources/performance/performance_hover/config/performance_hover.json
