+++ language-server/modules/langserver-core/src/test/resources/performance/performance_hover/source/performance_hover.bal
